Server

+++ client/src/components/DeleteTemplate.tsx
@@ -1,20 +1,37 @@
 import React from 'react'
 import { toast } from 'react-toastify';
 import TemplateClient from '../lib/templateClient';
+import { useGoogleReCaptcha } from 'react-google-recaptcha-v3';
+import { useAuthContext } from '../contexts/AuthContext';
 
 
 const DeleteTemplate = ({setOpenDelete, type, name, id, setReload}: {setOpenDelete: React.Dispatch<React.SetStateAction<boolean>>, type: 'email' | 'sms' | undefined, name: string, id: string, setReload: React.Dispatch<React.SetStateAction<boolean>>}) => {
     const [loader, setLoader] = React.useState(false);
+    const {executeRecaptcha} = useGoogleReCaptcha();
+    const {atk, setAtkFunc, logout} = useAuthContext();
     const handleDelete = async() => {
         setLoader(true);
+
+        if (!executeRecaptcha) {
+            toast.error("Execute recaptcha not yet available");
+            return;
+        }
+
+        const token = await executeRecaptcha('delete');
         try {
             const request = await TemplateClient.deleteTemplate({
                 type: type as 'email' | 'sms',
                 id
+            }, {
+                meta: {
+                    'x-recaptcha-token': token,
+                    'x-auth-token': atk
+                }
             });
             const response = request.response;
             if(response.status == true){
                 toast.success(response.message);
+                setAtkFunc(request.requestHeaders['x-auth-token'] as string);
                 setOpenDelete(false);
                 setReload(true);
                 return
@@ -22,6 +39,9 @@ const DeleteTemplate = ({setOpenDelete, type, name, id, setReload}: {setOpenDele
             toast.error(response.message);
         } catch (error) {
             toast.error(error instanceof Error ? error.message : "An unexpected error occurred.");
+            if(error instanceof Error && error.message=="unkown user"){
+                logout();
+            }
             if(import.meta.env.VITE_ENV === "development") console.error(error);
         }finally{
             setLoader(false);
